Rewriting System

It's just a struct that holds things together required for Knuth-Bendix completion.

KnuthBendix.RewritingSystem — Type

RewritingSystem{W<:AbstractWord, O<:Ordering}
RewritingSystem(rwrules::Vector{Pair{W,W}}, order, bare=false)

RewritingSystem holds the list of ordered (by order) rewriting rules of W<:AbstractWords.

KnuthBendix.rules — Method

rules(rws::RewritingSystem)

Return the iterator over active rewriting rules.

KnuthBendix.ordering — Method

ordering(rws::RewritingSystem)

Return the ordering of the rewriting system.

KnuthBendix.alphabet — Method

alphabet(rws::RewritingSystem)

Return the underlying Alphabet of the rewriting system.

KnuthBendix.isirreducible — Method

isirreducible(w::AbstractWord, rws::RewritingSystem)

Returns whether a word is irreducible with respect to a given rewriting system

KnuthBendix.irreduciblesubsystem — Method

irreduciblesubsystem(rws::RewritingSystem)

Return an array of left sides of rules from rewriting system of which all the proper subwords are irreducible with respect to this rewriting system.

KnuthBendix.check_confluence — Function

check_confluence(rws::RewritingSystem)

Check if a reduced rewriting system is confluent.

Return a stack of critical pairs. While the stack is by no means an exhaustive list of critical pairs, empty stack is returned if and only if rws is confluent.

There is also a modifying version check_confluence!.

KnuthBendix.isconfluent — Function

isconfluent(rws::RewritingSystem)

Check if a rewriting system is confluent.

The check follows first by reducing rws using KBS2AlgPlain(), and then constructing its index automaton and running backtrack search for all rules of the system.

Note

isconfluent may modify rws if it is not already reduced.

KnuthBendix.reduce! — Method

reduce!(rws::RewritingSystem[, work=Workspace(rws); kwargs...])

Reduce the rewriting system in-place using the default algorithm

Currently the default algorithm is KBS2AlgPlain(), see reduce!(::KBS2AlgPlain, ...).